feat: 사용자 API 수정
model.py 분리
에러 메시지 변경
