+++ sejongsql/app_main/models/__init__.py
@@ -1,4 +1,6 @@
 from .sample import SamplePost, SampleComment
-from .model import User, Class, UserBelongClass, ApiLog
+from .user import User
+from .classes import Class, UserBelongClass
 from .problem import ProblemGroup, Problem, UserSolveProblem, ExplainWarning, WarningMatchProblem, WarningMatchUp
 from .env import Env, TableBelongEnv, EnvBelongClass
+from .log import ApiLog
